+++ cloudfirestore/typed_batch_getter.go
@@ -1,0 +1,17 @@
+package cloudfirestore
+
+import (
+	"context"
+
+	"cloud.google.com/go/firestore"
+)
+
+type FuncGetDoc func(ids ...string) *firestore.DocumentRef
+
+type TypedBatchGetter[T any] interface {
+	Add(ids ...string)
+	Delete(ids ...string)
+	GetMap() map[string]*T
+	Get(ids ...string) *T
+	Commit(ctx context.Context) error
+}
